gcp_course/python/cloud_functions/zookeeper_backup/requirements.txt
-i https://pypi.org/simple
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
google-api-core==1.22.1
google-auth==1.21.0
google-cloud-core==1.4.1
google-cloud-storage==1.31.0
google-crc32c==1.0.0 ; python_version >= '3.5'
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
idna==2.10
kazoo==2.8.0
protobuf==3.13.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pytz==2020.1
requests==2.24.0
rsa==4.6 ; python_version >= '3.5'
six==1.15.0
urllib3==1.25.10
